+++ manheim_c7n_tools/config.py
@@ -116,10 +116,6 @@ class ManheimConfig(object):
         self.config_path = kwargs.pop('config_path')
         logger.debug('Validating configuration...')
         jsonschema.validate(kwargs, MANHEIM_CONFIG_SCHEMA)
-        if kwargs['regions'][0] != 'us-east-1':
-            raise RuntimeError(
-                'ERROR: the first configured region must be us-east-1'
-            )
         self._config = kwargs
         self._config['account_id'] = str(self._config['account_id'])
         if 'cleanup_notify' not in self._config: